autodiscovery.py: Change thermostat mode from 'heat' to 'auto'
Change to cause the Thermostat to report mode 'auto'. This is done
by setting auto as the only available mode, using the state topic
as mode state topic and providing a static mode state topic
enforcing state auto.

This way the thermostat is properly visualised and not assumed to be always
off.